[Chore/#41] AppPopUp view / viewController 분리
